requirements.txt

asgiref
certifi
distlib
dj-database-url
dj-rest-auth
Django
django-allauth
django-cors-headers
django-crispy-forms
django-email
django-environ
django-filter
djangorestframework
filelock
Markdown
MouseInfo
mysql-connector-python
pillow
pipenv
platformdirs
psycopg2
psycopg2-binary
PyAutoGUI
PyGetWindow
PyMsgBox
pyperclip
PyRect
PyScreeze
pytweening
setuptools
sqlparse
typing_extensions
tzdata
virtualenv